+++ backend/app/e_paul_smarthome_system/urls.py
@@ -1,6 +1,6 @@
 from django import urls
-from .views.rightsSettings import RightsSettings
-from .views.userView import UserView, SingleUserView, SingleUserSettingsView
+from .views.settings import RightsSettings, SingleUserSettingsView, ChangePin
+from .views.userView import UserView, SingleUserView
 from .views.accountView import AccountView
 from .views.about_us import AboutUsView
 from .views.login import Login, LoginUser
@@ -9,6 +9,7 @@ from .views.logout import Logout
 from .views.home import HomeView
 from .views.devices import DeviceView
 from .views.validate import ValidatePin
+from .views.getData import GetUser
 from django.urls import path
 
 
@@ -27,5 +28,7 @@ urlpatterns = [
     path("devices", DeviceView.as_view(), name = "devices"), 
     path("validatePin", ValidatePin.as_view(), name = "validatePin"),
     path("settings/rights", RightsSettings.as_view(), name = "rightsSettings"),
+    path("settings/pin", ChangePin.as_view(), name = "validatePin"),
     path("settings/<userid>", SingleUserSettingsView.as_view(), name = "userSettings"),
+    path("getUser/<accountId>", GetUser.as_view(), name = "getUser"),
 ]
